+++ TakeCommand/AssemblyVersion.cs
@@ -5,4 +5,4 @@
   
  using System.Reflection;
 
- [assembly: AssemblyVersion("1.4.12.0")]
+ [assembly: AssemblyVersion("1.4.12.1")]