+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+1.4.12.1
+	Added fix so returning to editor won't prevent spawning again
+	Added check for female kerbal 
+
 1.4.12
 	Updated for 1.4.1
 	Updated build scripts